# Speed up simulation by ~35 times by using Yosys and Verilator