--- a/HDL/sm83/Icarus/.gitignore
+++ b/HDL/sm83/Icarus/.gitignore
@@ -1,4 +1,5 @@
 *.run
 *.vcd
 out.mem
-obj_dir
+verilator_build
+verilator_pgo_build