… HDL/sm83/Icarus/roms/01-special.gb
… HDL/sm83/Icarus/roms/01-special.mem