@ogamespec 2024-09-15
I'm running PR, everything is fine there, thanks for the detailed descriptions you do, future generations will be grateful :)
I'll add from myself - I'm pleased that the results go to the emulator, since you probably noticed that this is one of the topics I'm interested in.
I also started to make SM83 simulator in C, small part of what I managed to do (decoder) is already in repository. But first I want to add all accumulated materials on SoC.